Improved split of green vip module.
Supports demand loading of green module rather than always loading it
with vip. There is also less magic involved.
